Split into separate files and add makefile
